feat.: Import UPENN_nda JSONs

+++ formsdb/runners/imports/export_upenn_mongo_to_psql.py
@@ -22,7 +22,7 @@ except ValueError:
 
 import logging
 from datetime import datetime
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Set
 
 from rich.logging import RichHandler
 
@@ -44,7 +44,7 @@ logargs = {
 logging.basicConfig(**logargs)
 
 
-def export_forms(config_file: Path):
+def export_forms(config_file: Path) -> None:
     """
     Export UPENN forms data from MongoDB to PostgreSQL.
 
@@ -55,58 +55,78 @@ def export_forms(config_file: Path):
         None
     """
     mongodb = db.get_mongo_db(config_file)
-    forms = mongodb["upenn"]
 
     forms_data: List[upenn_forms_model.UpennForms] = []
     sql_queries: List[str] = []
 
-    f_count = 0
-    for form in forms.find():
-        form_dict: Dict[str, Any] = form
-        subject_id = form_dict["_id"]
-        source_m_date = form_dict["_source_mdate"]
+    for collection in ["upenn_nda", "upenn"]:
+        logger.info(f"Exporting {collection} forms...")
+        forms = mongodb[collection]
+        f_count = 0
+        for form in forms.find():
+            form_dict: Dict[str, Any] = form
+            subject_id = form_dict["_id"]
+            source_m_date = form_dict["_source_mdate"]
 
-        for key, value in form_dict.items():
-            if isinstance(value, dict):
-                event_name = key
-                event_form_data = value
+            for key, value in form_dict.items():
+                if isinstance(value, dict):
+                    event_name = key
+                    event_form_data = value
 
-                for event_type, value in event_form_data.items():
-                    if isinstance(value, dict):
-                        event_type_form_data = value
+                    if collection == "upenn":
+                        for event_type, value in event_form_data.items():
+                            if isinstance(value, dict):
+                                event_type_form_data = value
 
-                        for key, value in event_type_form_data.items():
+                                for key, value in event_type_form_data.items():
+                                    if isinstance(value, datetime):
+                                        # convert datetime to isoformat
+                                        event_type_form_data[key] = value.isoformat()
+
+                                forms_data.append(
+                                    upenn_forms_model.UpennForms(
+                                        subject_id=subject_id,
+                                        event_name=event_name,
+                                        event_type=event_type,
+                                        form_data=event_type_form_data,
+                                        source_m_date=source_m_date,
+                                    )
+                                )
+                                f_count += 1
+                    elif collection == "upenn_nda":
+                        event_type = "nda"
+                        for key, value in event_form_data.items():
                             if isinstance(value, datetime):
                                 # convert datetime to isoformat
-                                event_type_form_data[key] = value.isoformat()
+                                event_form_data[key] = value.isoformat()
 
                         forms_data.append(
                             upenn_forms_model.UpennForms(
                                 subject_id=subject_id,
                                 event_name=event_name,
                                 event_type=event_type,
-                                form_data=event_type_form_data,
+                                form_data=event_form_data,
                                 source_m_date=source_m_date,
                             )
                         )
                         f_count += 1
-
-    logger.info(f"Total forms: {f_count}")
+        logger.info(f"Total forms: {f_count}")
 
     logger.info("Constructing SQL queries...")
+    purged_subjects: Set[str] = set()
     for form in forms_data:
         if not data.check_if_subject_exists(
             config_file=config_file, subject_id=form.subject_id
         ):
-            logger.warning(f"Subject {form.subject_id} already exists. Skipping...")
+            logger.warning(f"Subject {form.subject_id} does not exist in the database, skipping...")
             continue
 
-        sql_query = (
-            f"""DELETE FROM upenn_forms WHERE subject_id = '{form.subject_id}';"""
-        )
-
-        if sql_query not in sql_queries:
+        if form.subject_id not in purged_subjects:
+            sql_query = (
+                f"""DELETE FROM upenn_forms WHERE subject_id = '{form.subject_id}';"""
+            )
             sql_queries.append(sql_query)
+            purged_subjects.add(form.subject_id)
 
         sql_query = f"""
             INSERT INTO upenn_forms (subject_id, event_name, event_type,
@@ -118,7 +138,7 @@ def export_forms(config_file: Path):
         sql_queries.append(db.handle_null(sql_query))
 
     db.execute_queries(
-        config_file=config_file, queries=sql_queries, show_commands=False
+        config_file=config_file, queries=sql_queries, show_commands=False, show_progress=True
     )
